+++ CMakeLists.txt
@@ -45,7 +45,8 @@ else()
 endif()
 
 include_directories(include)
-add_library(coreforecast SHARED src/coreforecast.cpp)
+file(GLOB SOURCES src/*.cpp)
+add_library(coreforecast SHARED ${SOURCES})
 if(MSVC)
     set_target_properties(coreforecast PROPERTIES OUTPUT_NAME "libcoreforecast")
 endif()
