+++ include/export.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#ifdef _MSC_VER
+#define DLL_EXPORT __declspec(dllexport)
+#else
+#define DLL_EXPORT
+#endif
