+++ include/stats.h
@@ -1,0 +1,27 @@
+#pragma once
+
+#include <algorithm>
+
+template <typename T> inline T Quantile(T *data, T p, int n) {
+  T i_plus_g = p * (n - 1);
+  int i = static_cast<int>(i_plus_g);
+  T g = i_plus_g - i;
+  std::nth_element(data, data + i, data + n);
+  T out = data[i];
+  if (g > 0.0) {
+    std::nth_element(data, data + i + 1, data + n);
+    out += g * (data[i + 1] - out);
+  }
+  return out;
+}
+
+template <typename T> inline T SortedQuantile(T *data, T p, int n) {
+  T i_plus_g = p * (n - 1);
+  int i = static_cast<int>(i_plus_g);
+  T g = i_plus_g - i;
+  T out = data[i];
+  if (g > 0.0) {
+    out += g * (data[i + 1] - out);
+  }
+  return out;
+}
